% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Phenotypes.R
\name{getPhenotypeLog}
\alias{getPhenotypeLog}
\title{Get phenotype log}
\usage{
getPhenotypeLog(cohortIds = NULL, showHidden = FALSE)
}
\arguments{
\item{cohortIds}{IDs of cohorts to extraction from the library.}

\item{showHidden}{Some cohorts in the library are designed to be hidden. They are not
retrieved by default. To retrieve such cohorts, please set showHidden as TRUE.
Examples of hidden cohorts are withdrawn, deprecated, referrent cohorts.}
}
\value{
Returns a table with one row per cohort definitions with log information such as its release cycle.
Example, this function gives us insight on when a cohort definition was added/updated/deprecated
by the OHDSI PhenotypeLibrary.

A tibble.
}
\description{
Get phenotype log
}
\examples{
getPhenotypeLog(cohortIds = c(1, 2))

}
