% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Phenotypes.R
\name{getPlCohortDefinitionSet}
\alias{getPlCohortDefinitionSet}
\title{Get a cohort definition set}
\usage{
getPlCohortDefinitionSet(cohortIds)
}
\arguments{
\item{cohortIds}{IDs of cohorts to extraction from the library.}
}
\value{
A tibble with the cohort ID, name, sql, and JSON for the provided cohort IDs. Can be used by the
\code{CohortGenerator} package.
}
\description{
Get a cohort definition set
}
\examples{
cohorts <- listPhenotypes()
subsetIds <- cohorts$cohortId[1:3]
getPlCohortDefinitionSet(subsetIds)

}
