% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Phenotypes.R
\name{getPlConceptDefinitionSet}
\alias{getPlConceptDefinitionSet}
\title{Get conceptSets in cohorts}
\usage{
getPlConceptDefinitionSet(cohortIds = getPhenotypeLog()$cohortId)
}
\arguments{
\item{cohortIds}{IDs of cohorts to extraction from the library.}
}
\value{
Returns a table with one row per concept set for given cohort definitions.

A tibble.
}
\description{
Get conceptSets in cohorts
}
\examples{
cohorts <- getPhenotypeLog()
subsetIds <- cohorts$cohortId[1:3]
getPlConceptDefinitionSet(cohortIds = subsetIds)

}
