% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Phenotypes.R
\name{listPhenotypes}
\alias{listPhenotypes}
\title{List all phenotypes in the library}
\usage{
listPhenotypes()
}
\value{
A tibble with the cohort ID and name.
}
\description{
List all phenotypes in the library
}
\examples{
listPhenotypes()

}
