% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PhenotypeLibrary-package.R
\docType{package}
\name{PhenotypeLibrary-package}
\alias{PhenotypeLibrary}
\alias{PhenotypeLibrary-package}
\title{PhenotypeLibrary: A curated library of content generated by the OHDSI community on phenotypes and cohort definitions.}
\description{
OHDSI Phenotype Library is an open community resource maintained by the OHDSI community to support phenotype development, evaluation, sharing and re-use. The Phenotype Library is maintained by community librarians.
}
\author{
\strong{Maintainer}: Gowtham Rao \email{rao@ohdsi.com} (\href{https://orcid.org/0000-0002-4949-7236}{ORCID})

Authors:
\itemize{
  \item Martijn Schuemie \email{schuemie@ohdsi.org}
  \item Patrick Ryan
  \item Azza Shoaibi
}

Other contributors:
\itemize{
  \item Observational Health Data Science and Informatics [copyright holder]
}

}
\keyword{internal}
