% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Phenotypes.R
\name{updatePhenotypeLog}
\alias{updatePhenotypeLog}
\title{Update phenotype log}
\usage{
updatePhenotypeLog(updates)
}
\arguments{
\item{updates}{Data to update to the log. This is usually the output of ROhdsiWebApi::getCohortDefinitionsMetaData(baseUrl = baseUrl)}
}
\value{
Updates Phenotype Log related to added/updated/deprecated of the OHDSI PhenotypeLibrary.

A tibble.
}
\description{
Update phenotype log
}
