% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Phenotypes.R
\name{getPhenotypeLog}
\alias{getPhenotypeLog}
\title{Get phenotype log}
\usage{
getPhenotypeLog(cohortIds = NULL)
}
\arguments{
\item{cohortIds}{IDs of cohorts to extraction from the library.}
}
\value{
Returns a table with one row per cohort definitions with log information such as its release cycle.
Example, this function gives us insight on when a cohort definition was added/updated/deprecated
by the OHDSI PhenotypeLibrary.

A tibble.
}
\description{
Get phenotype log
}
\examples{
getPhenotypeLog(cohortIds = c(1, 2))

}
