% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Phenotypes.R
\name{listPhenotypes}
\alias{listPhenotypes}
\title{Deprecated. List all phenotypes in the library.}
\usage{
listPhenotypes()
}
\value{
A tibble with the cohort ID and name.
Deprecated. Please use getPhenotypeLog
}
\description{
Deprecated. List all phenotypes in the library.
}
\examples{
listPhenotypes()

}
