% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PhenotypeLibrary-package.R
\docType{package}
\name{PhenotypeLibrary-package}
\alias{PhenotypeLibrary}
\alias{PhenotypeLibrary-package}
\title{PhenotypeLibrary: A curated library of content generated by the OHDSI community on phenotypes and cohort definitions.}
\description{
A curated library generated by the OHDSI community on phenotypes and cohort definitions. The definitions are expressed in JSON and SQL (OHDSI SQL) instructions and are compatible with OHDSI analytic tools and OHDSI OMOP CDM v5.0+. To be in the library, the definitions must have gone through diagnostics using OHDSI tools such as CohortDiagnostics and PheValuator. Supporting content are in folders literature, notes and evaluation. Studies may be executed by importing the definitions into OHDSI tools and compiling study executable (see https://github.com/OHDSI-studies)
}
\author{
\strong{Maintainer}: Gowtham Rao \email{gowthamrao@gmail.com} (\href{https://orcid.org/0000-0002-4949-7236}{ORCID})

}
\keyword{internal}
