% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PhenotypeLibrary.R
\docType{package}
\name{PhenotypeLibrary-package}
\alias{PhenotypeLibrary}
\alias{PhenotypeLibrary-package}
\title{PhenotypeLibrary: The OHDSI Phenotype library}
\description{
A repository to store the content of the OHDSI Phenotype library.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://ohdsi.github.io/PhenotypeLibrary/}
  \item \url{https://github.com/OHDSI/PhenotypeLibrary}
  \item Report bugs at \url{https://github.com/OHDSI/PhenotypeLibrary/issues}
}

}
\keyword{internal}
